refactor: reorganize landing components and remove unused files
- Refactored the landing page structure by importing components directly from their respective files, improving clarity and maintainability.
- Deleted unused constants and hooks related to landing animations, streamlining the codebase.
- Updated integration category imports to reflect the new structure, ensuring consistency across components.
