+++ apps/www/src/components/landing/cards/3d-modeling-card.tsx
@@ -2,7 +2,7 @@
 
 import { useEffect, useRef, useState } from "react";
 
-import type { IntegrationCategory } from "../constants";
+import type { IntegrationCategory } from "../../../lib/animation/constants";
 
 interface ThreeDModelingCardProps {
   category: IntegrationCategory;
