--- apps/www/src/components/landing/index.ts
@@ -1,13 +1,0 @@
-// Components
-export { CenterCard } from "./center-card";
-export { GridLines } from "./grid-lines";
-export { IntegrationCategories } from "./integration-categories";
-export { PageHeader } from "./page-header";
-
-// Constants
-export {
-  integrationCategories,
-  CENTER_SIZE,
-  CENTER_START,
-  GRID_SIZE,
-} from "./constants";
